pyproject.toml
[tool.poetry]
name = "flask_webpack_turbolinks"
version = "0.1.0"
description = ""
authors = ["Cristian Rasch <cristianrasch@fastmail.fm>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
python-dotenv = "^0.15.0"
Flask-Static-Digest = "^0.1.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
honcho = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
